match(synth_xbox): wave-19 Mic/FilterCoeffs reconstruction (4 fns + 8 STL)
Reference-less is_stub from target asm + headers. Native 418/418 (synth_xbox excluded).
- DSP::LowpassCoefficients 0->97.2, HighpassCoefficients 0->98.7 (RBJ biquad cookbook)
- MicManagerXbox::AddRemoteMic 0->98.8 (GainEffect XAPO + RegisterRemoteTalker;
  branchless noChain select), Poll 0->87.9 (mic/chat/buffer loops, 0xDEADBEEFFACEF0 sentinel)
- 8 vector<ChatBuffer> STL template instantiations to 100% (free harvest incl. the named
  _M_insert_overflow_aux target)
Blocked: MicXbox::Poll (magic-constant pitch math), MicManagerXbox::Init (ODR refactor across 3 XAPO headers too invasive).

+++ src/xdk/xvh2/xvh2.h
@@ -9,8 +9,8 @@ struct IXHV2Engine { /* Size=0x4 */
     virtual DWORD Lock(UINT32);
     virtual DWORD StartLocalProcessingModes(UINT32, void **, UINT32);
     virtual DWORD StopLocalProcessingModes(UINT32, void **, UINT32);
-    virtual DWORD StartRemoteProcessingModes(UINT32, void **, UINT32);
-    virtual DWORD StopRemoteProcessingModes(UINT32, void **, UINT32);
+    virtual DWORD StartRemoteProcessingModes(UINT64, void **, UINT32);
+    virtual DWORD StopRemoteProcessingModes(UINT64, void **, UINT32);
     virtual DWORD SetMaxDecodePackets(UINT32);
     virtual DWORD RegisterLocalTalker(UINT32);
     virtual DWORD UnregisterLocalTalker(UINT32);
